Add a finder for GuessIt
GuessIt is just a python library, so we aren't exactly linking against
it. Plus it's a run-time dependency not a compile time, but it's kind of
nice to enfore these things. Otherwise I'm sure we will get a number of
complaints from people just trying it out from git.

The finder still does not encorporte version numbers.