Add a FileSystemTracker
This class is responsible for monitoring the file system for the list of
videos and informing us when the list of videos changes. Currently, it
uses Baloo to do so, but that can be changed later on. The user may not
have Baloo enabled or they might not have their videos indexed.