Processor: resume operations on startup
On startup call the consumers on each queue so that they can continue
processing any items that may be in the queue. This is the reason we
made every queue persistent.

It's not the most elegant solution. This whole awesome design pattern of
using persistent queues and consumers is falling apart a little bit.