MovieDbConsumer: Wait for init signal
The AysncJobConsumer now understands the concept of the being "ready"
and does not process any events when it is not ready. TheMovieDbApi
informs it of being ready when it gets the init signal.

This isn't perfect as libtmdbqt does not try to initialize itself again
when we connect to the network.